split data processing per dataset
